Move verifyBuild to separate file
