+++ node-src/tasks/verify/index.ts
@@ -1,228 +1,9 @@
-import { exitCodes, setExitCode } from '../../lib/setExitCode';
 import { createTask, transitionTo } from '../../lib/tasks';
-import { delay } from '../../lib/utilities';
-import { Context, Task } from '../../types';
+import { Context } from '../../types';
 import { endActivity, startActivity } from '../../ui/components/activity';
-import brokenStorybook from '../../ui/messages/errors/brokenStorybook';
-import listingStories from '../../ui/messages/info/listingStories';
-import storybookPublished from '../../ui/messages/info/storybookPublished';
-import turboSnapEnabled from '../../ui/messages/info/turboSnapEnabled';
-import buildLimited from '../../ui/messages/warnings/buildLimited';
-import paymentRequired from '../../ui/messages/warnings/paymentRequired';
-import snapshotQuotaReached from '../../ui/messages/warnings/snapshotQuotaReached';
-import turboSnapUnavailable from '../../ui/messages/warnings/turboSnapUnavailable';
-import {
-  awaitingUpgrades,
-  dryRun,
-  initial,
-  pending,
-  publishFailed,
-  runOnlyFiles,
-  runOnlyNames,
-  success,
-} from '../../ui/tasks/verify';
+import { dryRun, initial, pending } from '../../ui/tasks/verify';
 import { publishBuild } from './publishBuild';
-
-const StartedBuildQuery = `
-  query StartedBuildQuery($number: Int!) {
-    app {
-      build(number: $number) {
-        startedAt
-        failureReason
-        upgradeBuilds {
-          completedAt
-        }
-      }
-    }
-  }
-`;
-
-interface StartedBuildQueryResult {
-  app: {
-    build: {
-      startedAt?: number;
-      failureReason?: string;
-      upgradeBuilds?: {
-        completedAt?: number;
-      }[];
-    };
-  };
-}
-
-// these fields must be part of authorizedBuildFieldsViaAppCode in the public api
-const VerifyBuildQuery = `
-  query VerifyBuildQuery($number: Int!) {
-    app {
-      build(number: $number) {
-        id
-        number
-        status(legacy: false)
-        specCount
-        componentCount
-        testCount
-        changeCount
-        errorCount: testCount(statuses: [BROKEN])
-        actualTestCount: testCount(statuses: [IN_PROGRESS])
-        actualCaptureCount
-        inheritedCaptureCount
-        interactionTestFailuresCount
-        webUrl
-        browsers {
-          browser
-        }
-        features {
-          uiTests
-          uiReview
-          isReactNativeApp
-        }
-        autoAcceptChanges
-        turboSnapEnabled
-        wasLimited
-        app {
-          manageUrl
-          setupUrl
-          account {
-            exceededThreshold
-            paymentRequired
-            billingUrl
-          }
-          repository {
-            provider
-          }
-        }
-        tests {
-          spec {
-            name
-            component {
-              name
-              displayName
-            }
-          }
-          parameters {
-            viewport
-            viewportIsDefault
-          }
-          mode {
-            name
-          }
-        }
-      }
-    }
-  }
-`;
-
-interface VerifyBuildQueryResult {
-  app: {
-    build: Context['build'];
-  };
-}
-
-// TODO: refactor this function
-// eslint-disable-next-line complexity, max-statements
-export const verifyBuild = async (ctx: Context, task: Task) => {
-  const { client } = ctx;
-  const { list, onlyStoryNames, onlyStoryFiles = ctx.onlyStoryFiles } = ctx.options;
-  const { matchesBranch } = ctx.git;
-
-  // It's not possible to set both --only-changed and --only-story-files and/or --only-story-names
-  // onlyStoryFiles may be passed directly, or calculated via --only-changed
-  if (onlyStoryFiles) {
-    transitionTo(runOnlyFiles)(ctx, task);
-  }
-  if (onlyStoryNames) {
-    transitionTo(runOnlyNames)(ctx, task);
-  }
-
-  let timeoutStart = Date.now();
-  const waitForBuildToStart = async () => {
-    const { storybookUrl } = ctx;
-    const { number, reportToken } = ctx.announcedBuild;
-    const variables = { number };
-    const options = { headers: { Authorization: `Bearer ${reportToken}` } };
-
-    const {
-      app: { build },
-    } = await client.runQuery<StartedBuildQueryResult>(StartedBuildQuery, variables, options);
-
-    if (build.failureReason) {
-      ctx.log.warn(brokenStorybook(ctx, { failureReason: build.failureReason, storybookUrl }));
-      setExitCode(ctx, exitCodes.STORYBOOK_BROKEN, true);
-      throw new Error(publishFailed(ctx).output);
-    }
-
-    if (!build.startedAt) {
-      // Upgrade builds can take a long time to complete, so we can't apply a hard timeout yet,
-      // instead we only timeout on the actual build verification, after upgrades are complete.
-      if (build.upgradeBuilds?.some((upgrade) => !upgrade.completedAt)) {
-        task.output = awaitingUpgrades(ctx, build.upgradeBuilds).output;
-        timeoutStart = Date.now() + ctx.env.CHROMATIC_POLL_INTERVAL;
-      } else if (Date.now() - timeoutStart > ctx.env.STORYBOOK_VERIFY_TIMEOUT) {
-        setExitCode(ctx, exitCodes.VERIFICATION_TIMEOUT);
-        throw new Error('Build verification timed out');
-      }
-
-      await delay(ctx.env.CHROMATIC_POLL_INTERVAL);
-      await waitForBuildToStart();
-      return;
-    }
-
-    const {
-      app: { build: startedBuild },
-    } = await client.runQuery<VerifyBuildQueryResult>(VerifyBuildQuery, variables, options);
-    ctx.build = { ...ctx.announcedBuild, ...ctx.build, ...startedBuild };
-  };
-
-  await Promise.race([
-    waitForBuildToStart(),
-    new Promise((_, reject) =>
-      setTimeout(
-        reject,
-        ctx.env.CHROMATIC_UPGRADE_TIMEOUT,
-        new Error('Timed out waiting for upgrade builds to complete')
-      )
-    ),
-  ]);
-
-  ctx.isPublishOnly = !ctx.build.features?.uiReview && !ctx.build.features?.uiTests;
-
-  if (list && ctx.build.tests) {
-    ctx.log.info(listingStories(ctx.build.tests));
-  }
-
-  if (ctx.turboSnap) {
-    if (ctx.turboSnap.unavailable) {
-      ctx.log.warn(turboSnapUnavailable(ctx));
-    } else if (ctx.build.turboSnapEnabled) {
-      ctx.log.info(turboSnapEnabled(ctx));
-    }
-  }
-
-  if (ctx.build.wasLimited) {
-    const { account } = ctx.build.app;
-    if (account?.exceededThreshold) {
-      ctx.log.warn(snapshotQuotaReached(account));
-      setExitCode(ctx, exitCodes.ACCOUNT_QUOTA_REACHED, true);
-    } else if (account?.paymentRequired) {
-      ctx.log.warn(paymentRequired(account));
-      setExitCode(ctx, exitCodes.ACCOUNT_PAYMENT_REQUIRED, true);
-    } else {
-      // Future proofing for reasons we aren't aware of
-      if (account) ctx.log.warn(buildLimited(account));
-      setExitCode(ctx, exitCodes.BUILD_WAS_LIMITED, true);
-    }
-  }
-
-  if (ctx.build && ctx.storybookUrl) {
-    ctx.log.info(storybookPublished(ctx));
-  }
-
-  transitionTo(success, true)(ctx, task);
-
-  if (list || ctx.isPublishOnly || matchesBranch?.(ctx.options.exitOnceUploaded)) {
-    setExitCode(ctx, exitCodes.OK);
-    ctx.skipSnapshots = true;
-  }
-};
+import { verifyBuild } from './verifyBuild';
 
 /**
  * Sets up the Listr task for verifying the uploaded Storybook.
