+++ node-src/tasks/verify/verifyBuild.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 
 import { exitCodes } from '../../lib/setExitCode';
 import TestLogger from '../../lib/testLogger';
-import { verifyBuild } from './index';
+import { verifyBuild } from './verifyBuild';
 
 const environment = {
   CHROMATIC_POLL_INTERVAL: 10,
